refactor[#523] : entities/follow/api 타입 리팩토링
- 컨벤션에 맞춰 type 폴더 생성
